Escape urls from ruby-jmeter
I copied in a commit from someone else who had removed the calls to
URI.encode because that method doesn't exist in Ruby 3. The commit I
copied in didn't escape the urls properly though, it just removed the
method calls so they weren't escaped which caused issues when I tried to
add extracted variables to the url.